src/cerebras/modelzoo/data_preparation/data_curation/requirements.txt
# Core data processing and ML libraries
datatrove[io,processing]>=0.5.0
dask>=2025.5.1
dask-jobqueue>=0.9.0
distributed>=2025.5.1
pandas>=2.2.3
numpy>=2.0.0
scipy>=1.15.3
scikit-learn>=1.6.1
torch>=2.6.0
absl-py==2.1.0
protobuf==4.25.3
torchaudio>=2.2.2
torchdata>=0.7.1
torchtext>=0.17.2
torchvision>=0.17.2


# NLP and text processing
nltk>=3.9.1
--index-url https://pypi.org/simple
faiss-cpu==1.7.4
--index-url https://devpi.cerebras.aws/root/main/+simple/
kenlm>=0.3.0
xformers==0.0.29
transformers>=4.53.1
sentence-transformers>=5.0.0
tokenizers>=0.21.2
sentencepiece>=0.2.0
huggingface-hub>=0.31.4
datasets>=4.0.0
# kenlm @ git+https://github.com/kpu/kenlm.git@master#egg=kenlm

# Web scraping and data extraction
trafilatura>=1.11.0
FastWARC>=0.14.0
Resiliparse>=0.14.0
warcio>=1.7.5
beautifulsoup4>=4.13.4
bs4>=0.0.2
lxml>=5.4.0
lxml-html-clean>=0.4.2
inscriptis>=2.6.0
jusText>=3.0.2
htmldate>=1.9.3
courlan>=1.3.2
tldextract>=5.3.0
python-magic>=0.4.27
faust-cchardet>=2.1.19

# HTTP and async libraries
aiohttp>=3.12.14
aiohappyeyeballs>=2.6.1
aiosignal>=1.4.0
async-timeout>=5.0.1
requests>=2.32.4
requests-file>=2.1.0
urllib3>=2.5.0
certifi>=2025.7.14
charset-normalizer>=3.4.2
idna>=3.10

# Utilities and tools
tqdm>=4.67.1
click>=8.2.1
loguru>=0.7.3
wget>=3.2
fasteners>=0.19
pyahocorasick>=2.2.0
prettytable>=3.16.0
humanize>=4.12.3
dateparser>=1.2.2
python-dateutil>=2.9.0
tzlocal>=5.3.1

# Data formats and serialization
pyarrow>=20.0.0
pyarrow-hotfix>=0.7
orjson>=3.11.0
PyYAML>=6.0.2
msgpack>=1.1.1
fsspec>=2025.3.0
h2>=4.2.0
hpack>=4.1.0

# Type hints and validation
pydantic>=2.11.7
pydantic-core>=2.33.2
annotated-types>=0.7.0
typing-extensions>=4.14.1
typing-inspection>=0.4.1

# Core Python utilities
packaging>=25.0
filelock>=3.18.0
joblib>=1.5.1
six>=1.17.0
attrs>=25.3.0
deprecated>=1.2.18
executing>=2.2.0
asttokens>=2.4.1

# Visualization and dashboard
bokeh>=3.7.3
matplotlib>=3.10.3
seaborn>=0.13.2

# Compression and encoding
lz4>=4.3.2
zstandard>=0.23.0
brotli>=1.0.9
xxhash>=0.8.0
ftfy>=6.3.1

# System and network utilities
psutil>=5.9.0
cloudpickle>=3.1.1
dill>=0.3.7
multiprocess>=0.70.15
sortedcontainers>=2.4.0
toolz>=1.0.0
cytoolz>=1.0.1
partd>=1.4.2
locket>=1.0.0
tblib>=3.1.0
zict>=3.0.0

# Template and markup
Jinja2>=3.1.6
MarkupSafe>=3.0.2
soupsieve>=2.7

# Development tools
devtools>=0.12.2
Cython>=0.29.37
pybind11>=3.0.0

# Additional libraries
regex>=2024.11.6
babel>=2.17.0
fonttools>=4.59.0
kiwisolver>=1.4.8
cycler>=0.12.1
pyparsing>=3.2.3
pillow>=11.1.0
contourpy>=1.3.1
wcwidth>=0.2.13
colorama>=0.4.6
wrapt>=1.17.0
gmpy2>=2.2.1
mpmath>=1.3.0
sympy>=1.14.0
networkx>=3.5
narwhals>=1.46.0
threadpoolctl>=3.6.0
bottleneck>=1.4.2
numexpr>=2.10.1
tornado>=6.5.1
frozenlist>=1.7.0
multidict>=6.6.3
yarl>=1.20.1
propcache>=0.3.1
hyperframe>=6.1.0
cffi>=1.17.1
pycparser>=2.22
pygments>=2.19.2
pysocks>=1.7.1
pytz>=2025.2
zipp>=3.23.0
importlib-metadata>=8.7.0
xyzservices>=2025.4.0
opentelemetry-api>=1.34.1
safetensors>=0.5.3
tld>=0.13.1
dataclasses>=0.6
fasttext-numpy2-wheel>=0.9.2
